used blueprints instead of monolith
