+++ flaskblog/templates/home.html
@@ -10,7 +10,7 @@
         <div class="article-metadata">
           <a
             class="mr-2"
-            href="{{ url_for('user_posts', username=post.author.username) }}"
+            href="{{ url_for('users.user_posts', username=post.author.username) }}"
           >
             {{ post.author.username }}
           </a>
@@ -19,7 +19,7 @@
         <h2>
           <a
             class="article-title"
-            href="{{ url_for('post', post_id=post.id) }}"
+            href="{{ url_for('posts.post', post_id=post.id) }}"
           >
             {{ post.title }}
           </a>
@@ -33,14 +33,14 @@
       {% if posts.page == page_num %}
         <a
           class="btn btn-info mb-4"
-          href="{{ url_for('home', page=page_num) }}"
+          href="{{ url_for('main.home', page=page_num) }}"
         >
           {{ page_num }}
         </a>
       {% else %}
         <a
           class="btn btn-outline-info mb-4"
-          href="{{ url_for('home', page=page_num) }}"
+          href="{{ url_for('main.home', page=page_num) }}"
         >
           {{ page_num }}
         </a>
