+++ flaskblog/templates/login.html
@@ -38,7 +38,7 @@
         <div class="col-xs-12 mb-3">
           {{ form.submit(class="btn btn-outline-info") }}
           <small class="text-muted ml-2">
-            <a href="{{ url_for('reset_request') }}">Forgot Password?</a>
+            <a href="{{ url_for('users.reset_request') }}">Forgot Password?</a>
           </small>
         </div>
       </fieldset>
@@ -46,7 +46,7 @@
   </div>
   <div class="border-top pt-3">
     <small class="text-muted">
-      Need An Account? <a class="ml-2" href="{{ url_for('register') }}">Sign Up Now</a>
+      Need An Account? <a class="ml-2" href="{{ url_for('users.register') }}">Sign Up Now</a>
     </small>
   </div>
 {% endblock content %}
