+++ flaskblog/templates/post.html
@@ -9,7 +9,7 @@
       <div class="article-metadata">
         <a
           class="mr-2"
-          href="{{ url_for('user_posts', username=post.author.username) }}"
+          href="{{ url_for('users.user_posts', username=post.author.username) }}"
         >
           {{ post.author.username }}
         </a>
@@ -18,7 +18,7 @@
           <div>
             <a
               class="btn btn-secondary btn-sm mt-1 mb-1"
-              href="{{ url_for('update_post', post_id=post.id) }}"
+              href="{{ url_for('posts.update_post', post_id=post.id) }}"
             >
               Update
             </a>
@@ -47,7 +47,7 @@
         </div>
         <div class="modal-footer">
           <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-          <form action="{{ url_for('delete_post', post_id=post.id) }}" method="POST">
+          <form action="{{ url_for('posts.delete_post', post_id=post.id) }}" method="POST">
             <input class="btn btn-danger" type="submit" value="Delete">
           </form>
         </div>
