+++ flaskblog/templates/user_posts.html
@@ -11,7 +11,7 @@
         <div class="article-metadata">
           <a
             class="mr-2"
-            href="{{ url_for('user_posts', username=post.author.username) }}"
+            href="{{ url_for('users.user_posts', username=post.author.username) }}"
           >
             {{ post.author.username }}
           </a>
@@ -20,7 +20,7 @@
         <h2>
           <a
             class="article-title"
-            href="{{ url_for('post', post_id=post.id) }}"
+            href="{{ url_for('posts.post', post_id=post.id) }}"
           >
             {{ post.title }}
           </a>
@@ -34,14 +34,14 @@
       {% if posts.page == page_num %}
         <a
           class="btn btn-info mb-4"
-          href="{{ url_for('user_posts', username=user.username, page=page_num) }}"
+          href="{{ url_for('users.user_posts', username=user.username, page=page_num) }}"
         >
           {{ page_num }}
         </a>
       {% else %}
         <a
           class="btn btn-outline-info mb-4"
-          href="{{ url_for('user_posts', username=user.username, page=page_num) }}"
+          href="{{ url_for('users.user_posts', username=user.username, page=page_num) }}"
         >
           {{ page_num }}
         </a>
