+++ flaskblog/users/forms.py
@@ -1,22 +1,21 @@
 from flask_login import current_user
 from flask_wtf import FlaskForm
 from flask_wtf.file import (
-    FileAllowed,
     FileField,
+    FileAllowed,
 )
 from wtforms import (
-    BooleanField,
-    PasswordField,
     StringField,
+    PasswordField,
     SubmitField,
-    TextAreaField,
+    ValidationError,
+    BooleanField,
 )
 from wtforms.validators import (
     DataRequired,
+    Length,
     Email,
     EqualTo,
-    Length,
-    ValidationError,
 )
 
 from flaskblog.models import User
@@ -54,6 +53,7 @@ class RegistrationForm(FlaskForm):
             raise ValidationError(
                 'That email is taken. Please choose a different one'
             )
+
 
 class LoginForm(FlaskForm):
     email = StringField(
@@ -98,12 +98,6 @@ class UpdateAccountForm(FlaskForm):
                 raise ValidationError(
                     'That email is taken. Please choose a different one'
                 )
-
-
-class PostForm(FlaskForm):
-    title = StringField('Title', validators=[DataRequired()])
-    content = TextAreaField('Content', validators=[DataRequired()])
-    submit = SubmitField('Post')
 
 
 class RequestResetForm(FlaskForm):
